Convert batch value to decimal instead of integer.

+++ vouchers/models.py
@@ -6,6 +6,8 @@ from django.contrib.auth.models import User
 
 from utils import send_report
 
+from decimal import Decimal
+
 class Batch(models.Model):
 
     class Meta:
@@ -49,7 +51,7 @@ def send_generation_report(sender, **kwargs):
         'voucher_type': VOUCHER_TYPE_DICT[instance.voucher_type],
         'quantity': instance.quantity,
         'value': str(instance.value) + ' GHS',
-        'total_value': str(int(instance.value) * int(instance.quantity)) + ' GHS'
+        'total_value': str(Decimal(instance.value) * int(instance.quantity)) + ' GHS'
     }
 
     send_report(context)
